remove mater utils
