+++ mater/lib/src/lib.rs
@@ -17,7 +17,6 @@ mod file_reader;
 mod multicodec;
 mod stores;
 mod unixfs;
-mod utils;
 mod v1;
 mod v2;
 
@@ -27,7 +26,6 @@ pub use file_reader::CarExtractor;
 pub use ipld_core::cid::Cid;
 pub use multicodec::{DAG_PB_CODE, IDENTITY_CODE, RAW_CODE};
 pub use stores::{create_filestore, Blockstore, Config, FileBlockstore};
-pub use utils::stream_blocks_metadata;
 pub use v1::{BlockMetadata, Header as CarV1Header, Reader as CarV1Reader, Writer as CarV1Writer};
 pub use v2::{
     verify_cid, Characteristics, Header as CarV2Header, Index, IndexEntry, IndexSorted,
